+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -1,6 +1,10 @@
+## 0.6.15
+
+* Fixes incorrect camera preview rotation when capture orientation is locked.
+
 ## 0.6.14
 
-* Fixes incorrect camera preview rotation.
+* Fixes incorrect camera preview rotation when capture orientation is unlocked.
 
 ## 0.6.13